refactor: first batch of typescript-fication
