+++ package.json
@@ -23,6 +23,7 @@
     "zlib-sync": "^0.1.8"
   },
   "devDependencies": {
+    "@types/better-sqlite3": "^7.6.4",
     "typescript": "^5.1.6"
   }
 }
