--- src/database.js
@@ -1,6 +1,0 @@
-const Database = require('better-sqlite3');
-const db = new Database(__dirname + '/db.sqlite', {
-    fileMustExist: true
-});
-//db.defaultSafeIntegers(true);
-module.exports = db;
