+++ tsconfig.json
@@ -1,12 +1,13 @@
 {
 	"exclude": [ "dist", "node_modules" ],
-	"include": [ "src" ],
+	"include": [ "src", "src/**/*.json" ],
 	"compilerOptions": {
 		"module": "CommonJS",
 		"target": "ES6",
 		"skipLibCheck": true,
 		"strict": true,
 		"sourceMap": true,
+		"resolveJsonModule": true,
 
 		/* Paths */
 		"outDir": "./dist",
